Add JsonEncoder and JsonDecoder implmentations to their Bson counterparts (#1)
Kotlinx requires the encoder to implement `JsonEncoder` and `JsonDecoder` in order to encode generic `JsonElement` types.

The lack of this support for this came up many times when using [KBson](https://github.com/jershell/kbson) and [KMongo](https://github.com/Litote/kmongo) previously and I'm hoping we can quickly mitigate this now that MongoDB has official support.

I've added each implementation to support most types on decoding and determine the best possible type on encoding to insert. It is somewhat assumed that if someone is encoding a `JsonElement`, then they will probably be decoding one as well and, therefore, type information is primarily for storage optimization.

Unfortunately, Kotlinx writes all numbers as 64 bit integers first and never attempts to check for 32 bits or allow overriding the `JsonElementSerializer` behavior; this means that all integers will probably be written as `int64`. Regardless, this should be a valuable add.

* Add Kotlinx serialization JSON dependency

* Add implementation for JsonEncoder in DefaultBsonEncoder

* Add implementation for JsonDecoder in DefaultBsonDecoder

* Fix writing as an integer when the number is a decimal

* Add test for DataClassWithJsonElement

* Adjust encode and decode functions to comply with checks

---------

Signed-off-by: Mark <mark.vogel@nomadgcs.com>